refactor(skeleton-row): rename skeleton table to skeleton row
