+++ packages/forma-36-react-components/src/components/Skeleton/SkeletonRow/SkeletonRow.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { shallow, mount } from 'enzyme';
 import { axe } from 'jest-axe';
 
-import SkeletonTable from './SkeletonTable';
+import SkeletonRow from './SkeletonRow';
 import Table from '../../Table';
 import TableBody from '../../Table/TableBody';
 
@@ -10,7 +10,7 @@ it('renders the component', () => {
   const output = shallow(
     <Table>
       <TableBody>
-        <SkeletonTable />
+        <SkeletonRow />
       </TableBody>
     </Table>,
   );
@@ -22,7 +22,7 @@ it('renders the component with 10 cells', () => {
   const output = shallow(
     <Table>
       <TableBody>
-        <SkeletonTable columnCount={10} />
+        <SkeletonRow columnCount={10} />
       </TableBody>
     </Table>,
   );
@@ -34,7 +34,7 @@ it('has no a11y issues', async () => {
   const output = mount(
     <Table>
       <TableBody>
-        <SkeletonTable />
+        <SkeletonRow />
       </TableBody>
     </Table>,
   ).html();
