+++ packages/forma-36-react-components/src/components/Skeleton/SkeletonRow/SkeletonTableCell/index.ts
