+++ packages/forma-36-react-components/src/components/Skeleton/SkeletonRow/index.ts
@@ -1,0 +1,2 @@
+export * from './SkeletonRow';
+export { default } from './SkeletonRow';
