--- packages/forma-36-react-components/src/components/Skeleton/SkeletonTable/index.ts
@@ -1,2 +1,0 @@
-export * from './SkeletonTable';
-export { default } from './SkeletonTable';
